0337/rob_test.go
package _337

import "testing"

func Test_rob(t *testing.T) {
	type args struct {
		root *TreeNode
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{name: `输入: [3,2,3,null,3,null,1]
				
					 3
					/ \
				   2   3
					\   \ 
					 3   1
				
				输出: 7 
				解释: 小偷一晚能够盗取的最高金额 = 3 + 3 + 1 = 7.`,
			args: args{
				root: &TreeNode{
					Val: 3,
					Left: &TreeNode{
						Val:  2,
						Left: nil,
						Right: &TreeNode{
							Val:   3,
							Left:  nil,
							Right: nil,
						},
					},
					Right: &TreeNode{
						Val:  3,
						Left: nil,
						Right: &TreeNode{
							Val:   1,
							Left:  nil,
							Right: nil,
						},
					},
				}}, want: 7},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := rob(tt.args.root); got != tt.want {
				t.Errorf("rob() = %v, want %v", got, tt.want)
			}
		})
	}
}
